Remove mixin, swap `ConditionData`'s base class
